requirements.txt
asgiref==3.3.1
boto3==1.16.55
botocore==1.19.55
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
Django==3.1.5
django-cors-headers==3.6.0
django-macaddress==1.7.0
django-storages==1.11.1
django-user-agents==0.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
moviepy==1.0.3
netaddr==0.8.0
numpy==1.19.5
pandas==1.2.2
Pillow==8.1.0
proglog==0.1.9
PyJWT==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
razorpay==1.2.0
requests==2.25.1
s3transfer==0.3.4
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.56.0
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.2
user-agents==2.2.0
